+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 1.0.0
 - Enable negative `lengthAngle` to configure clockwise and counterclockwise charts
-
 - Make counterclockwise charts by default
+- Make <ReactMinimalPieChart> a pure component
 
 # 0.0.2
 - Incorporate `svg-partial-circle` lib